test(doctor): isolate workspace precedence regression
